--- eventmesh-common/src/main/java/org/apache/eventmesh/common/transform/Template.java
@@ -1,47 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.eventmesh.common.transform;
-
-import org.apache.eventmesh.common.exception.EventMeshException;
-
-import org.apache.commons.text.StringSubstitutor;
-
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class Template {
-
-    private String template;
-
-    public String substitute(List<Variable> variables) throws EventMeshException {
-
-        Map<String, String> valuesMap = variables.stream()
-            .filter(variable -> variable.getJsonPath() != null)
-            .collect(Collectors.toMap(Variable::getName, Variable::getJsonPath));
-        StringSubstitutor sub = new StringSubstitutor(valuesMap);
-
-        return sub.replace(template);
-
-    }
-}
