+++ eventmesh-transform/src/main/java/org/apache/eventmesh/transform/TemplateTransformer.java
@@ -1,0 +1,44 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.eventmesh.transform;
+
+import java.util.List;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+
+class TemplateTransformer implements Transformer {
+
+    private final JsonPathParser jsonPathParser;
+
+    private final Template template;
+
+    TemplateTransformer(JsonPathParser jsonPathParser, Template template) {
+        this.template = template;
+        this.jsonPathParser = jsonPathParser;
+    }
+
+    @Override
+    public String transform(String json) throws JsonProcessingException {
+        // 1: get variable match results
+        List<Variable> variableList = jsonPathParser.match(json);
+        // 2: use results replace template
+        String res = template.substitute(variableList);
+        return res;
+    }
+
+}
